+++ src/packagedcode/__init__.py
@@ -26,7 +26,6 @@ from packagedcode import freebsd
 from packagedcode import godeps
 from packagedcode import golang
 from packagedcode import haxe
-from packagedcode import lkm
 from packagedcode import maven
 from packagedcode import misc
 from packagedcode import npm
@@ -233,8 +232,6 @@ SYSTEM_PACKAGE_DATAFILE_HANDLERS = [
     debian.DebianInstalledFilelistHandler,
     debian.DebianInstalledMd5sumFilelistHandler,
     debian.DebianInstalledStatusDatabaseHandler,
-
-    lkm.LinuxKernelModuleHandler,
 
     rpm.RpmLicenseFilesHandler,
     rpm.RpmMarinerContainerManifestHandler,